# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Proxy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HistoryEntry {
  id: number;
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  favicon?: string | null;
  visitedAt: string;
}

export interface Bookmark {
  id: number;
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  favicon?: string | null;
  createdAt: string;
}

export interface BookmarkInput {
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  favicon?: string | null;
}

export interface SuccessResponse {
  success: boolean;
}