# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Web Proxy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: proxy
    description: Proxy operations
  - name: history
    description: Browsing history
  - name: bookmarks
    description: Bookmarked URLs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /history:
    get:
      operationId: listHistory
      tags: [history]
      summary: Get recent browsing history
      responses:
        "200":
          description: List of recent URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"
    delete:
      operationId: clearHistory
      tags: [history]
      summary: Clear all history
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List bookmarks
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Add a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"

  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    HistoryEntry:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        title:
          type: ["string", "null"]
        favicon:
          type: ["string", "null"]
        visitedAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - visitedAt

    Bookmark:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        title:
          type: ["string", "null"]
        favicon:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - createdAt

    BookmarkInput:
      type: object
      properties:
        url:
          type: string
        title:
          type: ["string", "null"]
        favicon:
          type: ["string", "null"]
      required:
        - url

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success