# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Proxy API
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  favicon?: string | null;
  createdAt: Date;
}