# lib/api-zod/src/generated/types/bookmarkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Proxy API
 * OpenAPI spec version: 0.1.0
 */

export interface BookmarkInput {
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  favicon?: string | null;
}