# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Proxy API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  favicon?: string | null;
  visitedAt: Date;
}